refactor: organize validation tests
