+++ tests/validation/validators/compare-fields-validation.spec.ts
@@ -1,5 +1,5 @@
-import { InvalidParamError } from '../../src/presentation/errors'
-import { CompareFieldsValidation } from '../../src/validation/validators'
+import { InvalidParamError } from '../../../src/presentation/errors'
+import { CompareFieldsValidation } from '../../../src/validation/validators'
 
 const makeSut = (): CompareFieldsValidation => new CompareFieldsValidation('field', 'fieldToCompare')
